Revert changes introduced in #990
There is an issue with the changes to prune the yarn 2 cache that, in certain situations, would cause all devDependencies to be reinstalled during the prune stage which is the opposite result of the intended fix.  This can cause slug sizes to increase for customers so this PR restores the previous cache clearing behavior.